+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## CHANGE LOG
 
+### v6.2.1
+
+2014-04-03 issue [#41](https://github.com/qiniu/ios-sdk/pull/41)
+
+- [#28] NSString qiniu category
+- [#38] 更新配置，引入Travis
+- [#39] bugfix 增加@optional 调用时的判断
+
 ### v6.2.0
 
 2014-03-04 issue [#37](https://github.com/qiniu/ios-sdk/pull/37)